random richtingen uit

+++ Main/Particle.pde
@@ -1,10 +1,10 @@
 class Particle{
   color Color;
-  int xPositie, yPositie;
+  float xPositie, yPositie;
   int particleWidth, particleHeight;
-  int directionX, directionY;
+  float directionX, directionY;
   
-  Particle(int xPositie, int yPositie, int particleWidth, int particleHeight, int directionX, int directionY, color Color){
+  Particle(int xPositie, int yPositie, int particleWidth, int particleHeight, float directionX, float directionY, color Color){
     this.xPositie = xPositie;
     this.yPositie = yPositie;
     this.particleWidth = particleWidth;
